DROOLS-4958: Tests
